refactor(models): adopt imports
