+++ package.json
@@ -52,6 +52,9 @@
     "vitest": "~0.32.0"
   },
   "optionalDependencies": {
+    "@esbuild/darwin-arm64": "^0.19.4",
+    "@nx/nx-darwin-arm64": "^16.9.1",
+    "@nx/nx-darwin-x64": "^16.9.1",
     "@nx/nx-linux-x64-gnu": "16.7.4"
   },
   "config": {
