+++ packages/cli/src/index.ts
@@ -1,13 +1,17 @@
 import { yargsCli } from './lib/cli';
-import { yargsGlobalOptionsDefinition } from './lib/options';
+import { options } from './lib/options';
 import { middlewares } from './lib/middlewares';
 import { commands } from './lib/commands';
+
+export { options } from './lib/options';
+export { middlewares } from './lib/middlewares';
+export { commands } from './lib/commands';
 
 export const cli = (args: string[]) =>
   yargsCli(args, {
     usageMessage: 'Code PushUp CLI',
     scriptName: 'code-pushup',
-    options: yargsGlobalOptionsDefinition(),
+    options,
     middlewares,
     commands,
   });
