--- packages/cli/src/lib/implementation/base-command-config.ts
@@ -1,11 +1,0 @@
-import {
-  globalCliArgsSchema,
-  refineCoreConfig,
-  unrefinedCoreConfigSchema,
-} from '@quality-metrics/models';
-import { z } from 'zod';
-
-export const commandBaseSchema = refineCoreConfig(
-  globalCliArgsSchema.merge(unrefinedCoreConfigSchema),
-);
-export type CommandBase = z.infer<typeof commandBaseSchema>;
